feat: create category function moved to external file
