+++ js/script.js
@@ -1,10 +1,8 @@
 import cards from '../data/cards.js';
 import createCard from  '../js/createCard.js'
+import {createCategory} from  '../js/createCard.js'
 
 let cardsGrid = document.querySelector('.cards');
-let rotateIcons = document.querySelectorAll('.rotate');
-
-console.log(cards[0]);
 
 document.querySelector('.form-check-input').onclick = function() {
     let trainSwitch = document.querySelector('.switch');
@@ -45,33 +43,7 @@ class BaseCard {
 }
 
 
-function createCategory(cardImage, title, index) {
-    let col = document.createElement('div');
-    col.classList.add('col', 'col-sm-6', 'col-lg-4', 'col-xl-3');
-    col.id = index;
 
-    let card = document.createElement('div');
-    card.classList.add('card','category');
-
-    let image = document.createElement('img');
-    image.classList.add('card-img-top');
-
-    image.src = './data/' + cardImage;
-
-    let cardBody = document.createElement('div');
-    cardBody.classList.add('card-body','english');
-    let cardTitle = document.createElement('h5');
-    cardTitle.classList.add('title')
-    cardTitle.innerHTML = title;
-    cardBody.append(cardTitle);
-
-    card.append(image);
-    card.append(cardBody);
-
-
-    col.append(card);
-    cardsGrid.append(col);
-}
 
 let categories = [];
 let arrayOfcards = [];
